add authorization & log route

+++ estudoAPI/DDDBasico/DDDBasico.Application/Command/Users/DeleteUserCommand.cs
@@ -2,6 +2,7 @@
 using DDDBasico.Domain.Interfaces;
 using DDDBasico.Domain.Interfaces.Services;
 using MediatR;
+using Microsoft.AspNetCore.Http;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -18,11 +19,14 @@ namespace DDDBasico.Application.Users.Command
 
         private readonly IRepositoryUser _repository;
         private readonly ITokenService _tokenService;
+        private readonly IHttpContextAccessor _httpContextAcessor;
 
-        public DeleteUserCommandHandler(IRepositoryUser repository, ITokenService tokenService)
+
+        public DeleteUserCommandHandler(IRepositoryUser repository, ITokenService tokenService, IHttpContextAccessor httpContextAcessor)
         {
             _repository = repository;
             _tokenService = tokenService;
+            _httpContextAcessor = httpContextAcessor;
         }
 
 
@@ -30,7 +34,8 @@ namespace DDDBasico.Application.Users.Command
         {
             try
             {
-               /* if(_tokenService.ReturnIdToken("!") != request.Id.ToString()) return await Task.FromResult("User deleted");*/
+                var token = _httpContextAcessor.HttpContext.Request.Headers["Authorization"].ToString().Split("Bearer");
+                if (_tokenService.ReturnIdToken(token[1].TrimStart()) != request.Id.ToString()) return "Not authorized";
                 var user = _repository.GetById(request.Id);
                 if (user == null) return await Task.FromResult("User not found");
                 _repository.Remove(user);
